refactor: edit profile banner
