+++ the-name-card/src/components/index.ts
@@ -8,3 +8,4 @@ export * from './Footer/Footer';
 export * from './WidgetItem/WidgetItem';
 export * from './WidgetItem/Items';
 export * from './Dropdown';
+export * from './EditProfileBanner/EditProfileBanner';
