+++ the-name-card/src/pages/EditProfile/EditProfile.scss
@@ -40,45 +40,6 @@
 	}
 }
 
-.banner {
-	width: 100%;
-	height: 10rem;
-
-	position: relative;
-
-	.banner-image {
-		height: 100%;
-		width: 100%;
-
-		object-fit: cover;
-	}
-
-	.banner-setting {
-		display: flex;
-		gap: $x-tiny-gap;
-		padding: $xx-tiny-gap $x-tiny-gap;
-
-		position: absolute;
-		right: 0;
-		bottom: 0;
-
-		transform: translate(0, 100%);
-
-		.banner-icon {
-			height: 1.5rem;
-			width: 100%;
-		}
-	}
-
-	.color-picker {
-		position: absolute;
-		left: 50%;
-		top: 3rem;
-		transform: translate(-50%, 0);
-		z-index: 1;
-	}
-}
-
 .wrapper {
 	height: 7rem;
 }
